Separate out some reusable code
