+++ iml-manager-cli/src/main.rs
@@ -2,63 +2,17 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
-use futures::Future;
-use iml_manager_cli::{
-    api_utils,
-    manager_cli_error::{
-        DurationParseError, ImlManagerCliError, RunStratagemCommandResult,
-        RunStratagemValidationError,
-    },
-};
-use iml_wire_types::{ApiList, Command, Host};
-use prettytable::{Row, Table};
-use spinners::{Spinner, Spinners};
-use std::{fmt::Display, process::exit};
+mod api_utils;
+mod display_utils;
+mod manager_cli_error;
+mod stratagem;
+
+use display_utils::{generate_table, start_spinner};
+use iml_manager_cli::api_utils::run_cmd;
+use iml_wire_types::{ApiList, EndpointName, Host};
+use std::process::exit;
+use stratagem::stratagem_cli;
 use structopt::StructOpt;
-
-#[derive(Debug, StructOpt)]
-pub enum StratagemCommand {
-    /// Kickoff a Stratagem scan
-    #[structopt(name = "scan")]
-    Scan {
-        /// The name of the filesystem to scan
-        #[structopt(short = "f", long = "filesystem")]
-        fs: String,
-        /// The report duration
-        #[structopt(short = "r", long = "report", parse(try_from_str = "parse_duration"))]
-        rd: Option<u64>,
-        /// The purge duration
-        #[structopt(short = "p", long = "purge", parse(try_from_str = "parse_duration"))]
-        pd: Option<u64>,
-    },
-}
-
-#[derive(serde::Serialize, serde::Deserialize, Debug)]
-struct StratagemCommandData {
-    filesystem: String,
-    report_duration: Option<u64>,
-    purge_duration: Option<u64>,
-}
-
-fn parse_duration(src: &str) -> Result<u64, ImlManagerCliError> {
-    if src.len() < 2 {
-        return Err(DurationParseError::InvalidValue.into());
-    }
-
-    let mut val = String::from(src);
-    let unit = val.pop();
-
-    let val = val.parse::<u64>()?;
-
-    match unit {
-        Some('h') => Ok(val * 3_600_000),
-        Some('d') => Ok(val * 86_400_000),
-        Some('m') => Ok(val * 60_000),
-        Some('s') => Ok(val * 1_000),
-        Some('1'...'9') => Err(DurationParseError::NoUnit.into()),
-        _ => Err(DurationParseError::InvalidUnit.into()),
-    }
-}
 
 #[derive(Debug, StructOpt)]
 pub enum ServerCommand {
@@ -69,13 +23,14 @@ pub enum ServerCommand {
 
 #[derive(StructOpt, Debug)]
 #[structopt(name = "iml")]
+#[structopt(raw(setting = "structopt::clap::AppSettings::ColoredHelp"))]
 /// The Integrated Manager for Lustre Agent CLI
 pub enum App {
     #[structopt(name = "stratagem")]
     /// Work with Stratagem server
     Stratagem {
         #[structopt(subcommand)]
-        command: StratagemCommand,
+        command: stratagem::StratagemCommand,
     },
     #[structopt(name = "server")]
     /// Work with Storage Servers
@@ -83,84 +38,6 @@ pub enum App {
         #[structopt(subcommand)]
         command: ServerCommand,
     },
-}
-
-/// Takes an asynchronous computation (Future), runs it to completion
-/// and returns the result.
-///
-/// Even though the action is asynchronous, this fn will block until
-/// the future resolves.
-fn run_cmd<R: Send + 'static, E: Send + 'static>(
-    fut: impl Future<Item = R, Error = E> + Send + 'static,
-) -> std::result::Result<R, E> {
-    tokio::runtime::Runtime::new().unwrap().block_on_all(fut)
-}
-
-fn generate_table<Rows, R>(columns: &[&str], rows: Rows) -> Table
-where
-    R: IntoIterator,
-    R::Item: ToString,
-    Rows: IntoIterator<Item = R>,
-{
-    let mut table = Table::new();
-
-    table.add_row(Row::from(columns));
-
-    for r in rows {
-        table.add_row(Row::from(r));
-    }
-
-    table
-}
-
-fn start_spinner(msg: &str) -> impl FnOnce(Option<String>) -> () {
-    let grey = termion::color::Fg(termion::color::LightBlack);
-    let reset = termion::color::Fg(termion::color::Reset);
-
-    let s = format!("{}{}{}", grey, reset, msg);
-    let s_len = s.len();
-
-    let sp = Spinner::new(Spinners::Dots9, s);
-
-    move |msg_opt| match msg_opt {
-        Some(msg) => {
-            sp.message(msg);
-        }
-        None => {
-            sp.stop();
-            print!("{}", termion::clear::CurrentLine);
-            print!("{}", termion::cursor::Left(s_len as u16));
-        }
-    }
-}
-
-fn display_cmd_state(cmd: &Command) {
-    if cmd.errored {
-        display_error(format!("{} errored", cmd.message));
-    } else if cmd.cancelled {
-        println!("🚫 {} cancelled", cmd.message);
-    } else if cmd.complete {
-        display_success(format!("{} successful", cmd.message));
-    }
-}
-
-fn display_success(message: impl Display) {
-    let green = termion::color::Fg(termion::color::Green);
-    let reset = termion::color::Fg(termion::color::Reset);
-
-    println!("{}✔{} {}", green, reset, message);
-}
-
-fn display_error(message: impl Display) {
-    let red = termion::color::Fg(termion::color::Red);
-    let reset = termion::color::Fg(termion::color::Reset);
-
-    println!("{}✗{} {}", red, reset, message);
-}
-
-#[derive(serde::Serialize, serde::Deserialize, Debug)]
-struct CmdWrapper {
-    command: Command,
 }
 
 fn main() {
@@ -173,69 +50,7 @@ fn main() {
     log::debug!("Matching args {:?}", matches);
 
     match matches {
-        App::Stratagem { command } => match command {
-            StratagemCommand::Scan { fs, rd, pd } => {
-                let fut = {
-                    let client =
-                        iml_manager_client::get_client().expect("Could not create API client");
-                    iml_manager_client::post(
-                        client,
-                        "run_stratagem",
-                        serde_json::json!(StratagemCommandData {
-                            filesystem: fs,
-                            report_duration: rd,
-                            purge_duration: pd
-                        }),
-                    )
-                    .and_then(|resp| iml_manager_client::concat_body(resp))
-                    .from_err()
-                    .and_then(|(resp, body)| {
-                        let status = resp.status();
-                        if status.is_success() {
-                            Ok(serde_json::from_slice::<CmdWrapper>(&body)?)
-                        } else if status.is_client_error() {
-                            serde_json::from_slice::<RunStratagemValidationError>(&body)
-                                .map(std::convert::Into::into)
-                                .map(Err)?
-                        } else if status.is_server_error() {
-                            Err(RunStratagemValidationError {
-                                code: RunStratagemCommandResult::ServerError,
-                                message: "Internal server error.".into(),
-                            }
-                            .into())
-                        } else {
-                            Err(RunStratagemValidationError {
-                                code: RunStratagemCommandResult::UnknownError,
-                                message: "Unknown error.".into(),
-                            }
-                            .into())
-                        }
-                    })
-                };
-
-                let command: Result<CmdWrapper, ImlManagerCliError> = run_cmd(fut);
-
-                match command {
-                    Ok(CmdWrapper { command }) => {
-                        let stop_spinner = start_spinner(&command.message);
-
-                        let command = run_cmd(api_utils::wait_for_cmd(command))
-                            .expect("Could not poll command");
-
-                        stop_spinner(None);
-
-                        display_cmd_state(&command);
-
-                        exit(exitcode::OK)
-                    }
-                    Err(validation_error) => {
-                        display_error(validation_error);
-
-                        exit(exitcode::CANTCREAT)
-                    }
-                }
-            }
-        },
+        App::Stratagem { command } => stratagem_cli(command),
         App::Server { command } => match command {
             ServerCommand::List => {
                 let stop_spinner = start_spinner("Running command...");
@@ -243,7 +58,7 @@ fn main() {
                 let fut = {
                     let client =
                         iml_manager_client::get_client().expect("Could not create API client");
-                    iml_manager_client::get(client, "host", vec![("limit", 0)])
+                    iml_manager_client::get(client, Host::endpoint_name(), vec![("limit", 0)])
                 };
 
                 let result: Result<ApiList<Host>, _> = run_cmd(fut);
@@ -278,80 +93,5 @@ fn main() {
                 }
             }
         },
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_parse_duration_with_days() {
-        match parse_duration("273d") {
-            Ok(x) => assert_eq!(x, 23_587_200_000),
-            Err(_) => panic!("Duration parser should not have errored!"),
-        }
-    }
-
-    #[test]
-    fn test_parse_duration_with_hours() {
-        match parse_duration("273h") {
-            Ok(x) => assert_eq!(x, 982_800_000),
-            Err(_) => panic!("Duration parser should not have errored!"),
-        }
-    }
-
-    #[test]
-    fn test_parse_duration_with_minutes() {
-        match parse_duration("273m") {
-            Ok(x) => assert_eq!(x, 16_380_000),
-            Err(_) => panic!("Duration parser should not have errored!"),
-        }
-    }
-
-    #[test]
-    fn test_parse_duration_with_seconds() {
-        match parse_duration("273s") {
-            Ok(x) => assert_eq!(x, 273_000),
-            Err(_) => panic!("Duration parser should not have errored!"),
-        }
-    }
-
-    #[test]
-    fn test_parse_duration_with_invalid_unit() {
-        match parse_duration("273x") {
-            Ok(_) => panic!("Duration should have Errored with InvalidUnit"),
-            Err(e) => assert_eq!(
-                e.to_string(),
-                "Invalid unit. Valid units include 'h' and 'd'."
-            ),
-        }
-    }
-
-    #[test]
-    fn test_parse_duration_without_unit() {
-        match parse_duration("273") {
-            Ok(_) => panic!("Duration should have Errored with NoUnit"),
-            Err(e) => assert_eq!(e.to_string(), "No unit specified."),
-        }
-    }
-
-    #[test]
-    fn test_parse_duration_with_insufficient_data() {
-        match parse_duration("2") {
-            Ok(_) => panic!("Duration should have Errored with InvalidValue"),
-            Err(e) => assert_eq!(
-                e.to_string(),
-                "Invalid value specified. Must be a valid integer."
-            ),
-        }
-    }
-
-    #[test]
-    fn test_parse_duration_with_invalid_data() {
-        match parse_duration("abch") {
-            Ok(_) => panic!("Duration should have Errored with invalid digit"),
-            Err(e) => assert_eq!(e.to_string(), "invalid digit found in string"),
-        }
     }
 }
