+++ iml-manager-cli/src/manager_cli_error.rs
@@ -58,6 +58,7 @@ pub enum ImlManagerCliError {
     ParseDurationError(DurationParseError),
     RunStratagemValidationError(RunStratagemValidationError),
     SerdeJsonError(serde_json::error::Error),
+    DoesNotExist(&'static str),
 }
 
 impl std::fmt::Display for ImlManagerCliError {
@@ -69,6 +70,7 @@ impl std::fmt::Display for ImlManagerCliError {
             ImlManagerCliError::ParseDurationError(ref err) => write!(f, "{}", err),
             ImlManagerCliError::RunStratagemValidationError(ref err) => write!(f, "{}", err),
             ImlManagerCliError::SerdeJsonError(ref err) => write!(f, "{}", err),
+            ImlManagerCliError::DoesNotExist(ref err) => write!(f, "{} does not exist", err),
         }
     }
 }
@@ -88,6 +90,7 @@ impl std::error::Error for ImlManagerCliError {
             ImlManagerCliError::ParseDurationError(ref err) => Some(err),
             ImlManagerCliError::RunStratagemValidationError(ref err) => Some(err),
             ImlManagerCliError::SerdeJsonError(ref err) => Some(err),
+            ImlManagerCliError::DoesNotExist(_) => None,
         }
     }
 }
